update calendar page composition separate calendar and schedule components and pass props add schedule action btns
